+++ app/lang/en.php
@@ -1,0 +1,44 @@
+<?php
+/* Language: English */
+
+//define("LANG_", "");
+
+define("LANG_title", "DMS");
+define("LANG_footer", "DMS (Document Management System) by Bastiaan de hart");
+
+define("LANG_NAV_dashboard", "Dashboard");
+define("LANG_NAV_notes", "Notes");
+define("LANG_NAV_profile", "Profile");
+define("LANG_NAV_password", "Change password");
+define("LANG_NAV_sign_out", "Sign out");
+
+define("LANG_SIGNIN_title", "Please sign in");
+define("LANG_SIGNIN_username", "Email address");
+define("LANG_SIGNIN_password", "Password");
+define("LANG_SIGNIN_button", "Sign in");
+define("LANG_SIGNIN_error", "Username or password is incorrect.");
+
+define("LANG_NOTES_TABLE_heading", "All notes:");
+define("LANG_NOTES_TABLE_btn_new", "Make new note");
+define("LANG_NOTES_TABLE_title", "Title");
+define("LANG_NOTES_TABLE_lastedit", "Last edited");
+define("LANG_NOTES_TABLE_actions", "Actions");
+
+define("LANG_NOTES_TABLE_BTN_edit", "Edit");
+define("LANG_NOTES_TABLE_BTN_remove", "Remove");
+
+define("LANG_NOTES_EDIT_heading", "Edit note");
+
+define("LANG_NOTES_NEW_heading", "New note");
+define("LANG_NOTES_NEW_title", "Title");
+define("LANG_NOTES_NEW_note", "Note");
+define("LANG_NOTES_NEW_BTN_save", "Save");
+
+define("LANG_PROFILE_PASSWORD_heading", "Password change");
+define("LANG_PROFILE_PASSWORD_new_password", "New password right here");
+define("LANG_PROFILE_PASSWORD_new_password_repeat", "New password right here (repeat)");
+define("LANG_PROFILE_PASSWORD_new_password_btn", "Save");
+
+define("LANG_404_heading", "404 Not found");
+define("LANG_404_paragraph", "Oops! Something is broken.");
+?>