+++ app/lang/nl.php
@@ -1,0 +1,44 @@
+<?php
+/* Language: Dutch */
+
+//define("LANG_", "");
+
+define("LANG_title", "DMS");
+define("LANG_footer", "DMS (Document Management System) door Bastiaan de hart");
+
+define("LANG_NAV_dashboard", "Dashboard");
+define("LANG_NAV_notes", "Notities");
+define("LANG_NAV_profile", "Profiel");
+define("LANG_NAV_password", "Wachtwoord wijziggen");
+define("LANG_NAV_sign_out", "Uitloggen");
+
+define("LANG_SIGNIN_title", "Inloggen a.u.b.");
+define("LANG_SIGNIN_username", "Email address");
+define("LANG_SIGNIN_password", "Wachtwoord");
+define("LANG_SIGNIN_button", "Inloggen");
+define("LANG_SIGNIN_error", "Gebruikersnaam of wachtwoord onjuist.");
+
+define("LANG_NOTES_TABLE_heading", "Alle notities:");
+define("LANG_NOTES_TABLE_btn_new", "Nieuwe notitie");
+define("LANG_NOTES_TABLE_title", "Titel");
+define("LANG_NOTES_TABLE_lastedit", "Laatst bewerkt");
+define("LANG_NOTES_TABLE_actions", "Acties");
+
+define("LANG_NOTES_TABLE_BTN_edit", "Bewerken");
+define("LANG_NOTES_TABLE_BTN_remove", "Verwijderen");
+
+define("LANG_NOTES_EDIT_heading", "Bewerk notitie");
+
+define("LANG_NOTES_NEW_heading", "Nieuwe notitie");
+define("LANG_NOTES_NEW_title", "Titel");
+define("LANG_NOTES_NEW_note", "Notitie");
+define("LANG_NOTES_NEW_BTN_save", "Opslaam");
+
+define("LANG_PROFILE_PASSWORD_heading", "Wachtwoord wijziggen");
+define("LANG_PROFILE_PASSWORD_new_password", "Nieuw wachtwoord");
+define("LANG_PROFILE_PASSWORD_new_password_repeat", "Nieuw wachtwoord (repeat)");
+define("LANG_PROFILE_PASSWORD_new_password_btn", "Opslaan");
+
+define("LANG_404_heading", "404 Niet gevonden");
+define("LANG_404_paragraph", "Oops! Er is iets kappot.");
+?>