+++ app/views/global/head.view.php
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <html lang="<?php echo $template->getLanguage(); ?>">
-<head><?php echo $template->loadHead("css/app", "DMS"); ?></head>
+<head><?php echo $template->loadHead("css/app", LANG_title); ?></head>
 <body>
 <?php echo $template->loadheader(); ?>
 <main role="main" class="container">