+++ app/views/notes/edit.view.php
@@ -1,2 +1,2 @@
-<?php echo $template->addTitle("Edit note"); ?>
+<?php echo $template->addTitle(LANG_NOTES_EDIT_heading); ?>
 <?php echo $notes->editNoteForm($_GET["id"]); ?>