+++ app/views/notes/new.view.php
@@ -1,2 +1,2 @@
-<?php echo $template->addTitle("New note"); ?>
+<?php echo $template->addTitle(LANG_NOTES_NEW_heading); ?>
 <?php echo $notes->makeNoteForm(); ?>