+++ app/views/notes/view.view.php
@@ -1,3 +1,3 @@
-<?php echo $template->addTitle("Notes"); ?>
-<?php echo $template->addButton("index.php?page=notes&action=new", "Make new note"); ?>
+<?php echo $template->addTitle(LANG_NOTES_TABLE_heading); ?>
+<?php echo $template->addButton("index.php?page=notes&action=new", LANG_NOTES_TABLE_btn_new); ?>
 <?php echo $notes->viewNotes(); ?>