+++ app/views/profile/password.view.php
@@ -1,2 +1,2 @@
-<?php echo $template->addTitle("Password Change"); ?>
+<?php echo $template->addTitle(LANG_PROFILE_PASSWORD_heading); ?>
 <?php echo $profile->passwordChangeForm(); ?>