+++ app/views/user/sign-in.view.php
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
 <html lang="<?php echo $template->getLanguage(); ?>">
-<head><?php echo $template->loadHead("css/login", "DMS | Sign-In"); ?></head>
+<head><?php echo $template->loadHead("css/login", LANG_title); ?></head>
 <body><body class="text-center"><?php echo $template->loadLoginForm(); ?></body></body>
 </html>