+++ autoload.php
@@ -6,8 +6,6 @@ $mysqlUsername      = "root";
 $mysqlPassword      = "root";
 $mysqlDatabase      = "notes";
 
-$language           = "en";
-
 if ($debug === true) {
     ini_set('display_errors', 1);
     ini_set('display_startup_errors', 1);
@@ -17,7 +15,7 @@ if ($debug === true) {
 foreach (glob("app/class/*.php") as $file) { include $file; }
 
 $mysql          = new mysqli($mysqlHost, $mysqlUsername, $mysqlPassword, $mysqlDatabase);
-$template       = new Template("app/assets", $language);
+$template       = new Template("app/assets", "en");
 $views          = new Views();
 $controller     = new Controller();
 $profile        = new Profile();
@@ -25,6 +23,7 @@ $notes          = new Notes();
 $profile        = new Profile();
 $user           = new User();
 $bootstrap      = new Bootstrap();
-$dateTime       = new DT($language);
+$dateTime       = new DT("en");
 $b64            = new Base64();
+$lang           = new Language("en");
 ?>